Rootfs: Fix CROSS_COMPILE symbol and other problems
Turns out that CROSS_COMPILE is used by busybox build (and, as it happens,
by the kernel build), so restore the export prefix on this symbol.

Also fix local build by setting CC when invoking the build-local target.

Finally, generate a copy of the image file with IMAGE_NAME for convenience.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/rootfs@73324 e099a375-04f9-0310-9d5f-a741eaff62e1