Rootfs: Change use of fakeroot to enable build on RHEL6
Fakeroot cannot be used for programs built against a different
architecture, in particular a 64-bit fakeroot will not work properly
with a 32-bit cross compiler.  To work around this, reduce the amount
of work done inside fakeroot.

In this commit fakeroot is now only used where root is really necessary
and for the final cpio assembly.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/rootfs@77287 e099a375-04f9-0310-9d5f-a741eaff62e1