Manage build placement for ntp and screen packages
It turns out that cross compiling can take an uncomfortable number of cues
from the build environment.  screen uses the location of an installed copy
of screen to determine the default prefix ... and ntp does some even deeper
poking about on the host to determine where to place its files.

For these two packages the fixes are straightforward, but I have the
uncomfortable feeling this is the tip of an iceberg.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/targetOS/rootfs@158979 e099a375-04f9-0310-9d5f-a741eaff62e1